V.5.1
Fonctionnel

bugs : Rebond sur buzzer et le circuit se reset sans raisons rarement

+++ 2024_PO_CentralBoard.X/ComXLR8.c
@@ -1,6 +1,6 @@
 #include "ComXLR8.h"
 
-void startXLR8(int startBonus)
+void startXLR8(int8 startBonus)
 {
     if(startBonus == 1)
     {
